refactor: AppConfig 클래스명 변경
